New Order Cancellation features:
* Orders can be cancelled only if all the payments on that order are $0.
* Flex Pass Orders can only be cancelled if the flex pass in question hasn't been used to order any tickets
* Flex Pass Orders now cascade delete flex pass line items and flex passes when deleted

+++ app/models/orders/order.rb
@@ -406,8 +406,15 @@ class Order < ActiveRecord::Base
   end
 
   def cancel!
-    self.allow_deletion!
-    self.destroy
+
+    if self.payments.select{|p| !p.can_cancel?}.size > 0
+      errors.add(:error, "Cannot cancel orders with payments")
+      false
+    else
+      self.allow_deletion!
+      self.destroy
+      true
+    end
   end
 
   def allow_deletion!
